+++ example_battery.py
@@ -13,7 +13,12 @@ data_logger: DataLogger = DataLogger(config)
 
 def on_data_received(client: BatteryClient, data):
     logging.debug("{} => {}".format(client.device.alias(), data))
-    client.disconnect()
+    if config['remote_logging'].getboolean('enabled'):
+        data_logger.log_remote(json_data=data)
+    if config['mqtt'].getboolean('enabled'):
+        data_logger.log_mqtt(json_data=data)
+    if not config['device'].getboolean('enable_polling'):
+        client.disconnect()
 
 logging.info(f"Starting client: {config['device']['alias']} => {config['device']['mac_addr']}")
