+++ renogybt/DataLogger.py
@@ -28,7 +28,7 @@ class DataLogger:
             auth=auth,
             client_id="renogy-bt"
         )
-    
+
     def log_pvoutput(self, json_data):
         date_time = datetime.now().strftime("d=%Y%m%d&t=%H:%M")
         data = f"{date_time}&v1={json_data['power_generation_today']}&v2={json_data['pv_power']}&v3={json_data['power_consumption_today']}&v4={json_data['load_power']}&v5={json_data['controller_temperature']}&v6={json_data['battery_voltage']}"
