+++ renogybt/__init__.py
@@ -1,3 +1,3 @@
-from .BTOneClient import BTOneClient
+from .BTModuleClient import BTModuleClient
 from .DataLogger import DataLogger
 from .BatteryClient import BatteryClient
